Disclaimer
So far I have just added a link to the web page and changed the
disclaimer to be the first thing shown on the screen when opened. I will
begin to work on the rest of the pages including continue and back
buttons on most or all of the pages. Finally, I will create a rough
outline of all pages. It would be great if you started on coding the
inputs into the table to receive a final result (recommended or
optional)